pyproject.toml
[project]
name = "vaultlight-secure-notes-api"
version = "1.0.0"
description = "Secure notes REST API with JWT auth and encryption-at-rest."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.27.0",
  "pydantic>=2.6.0",
  "pydantic-settings>=2.2.0",
  "sqlalchemy>=2.0.0",
  "passlib[bcrypt]>=1.7.4",
  "python-jose>=3.3.0",
  "cryptography>=42.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "httpx>=0.27.0",
  "ruff>=0.6.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]